[transport] Move compression traits into a separate header
