+++ BUILD
@@ -1513,6 +1513,7 @@ grpc_cc_library(
         "//src/core:map",
         "//src/core:match",
         "//src/core:memory_quota",
+        "//src/core:metadata_compression_traits",
         "//src/core:no_destruct",
         "//src/core:notification",
         "//src/core:packed_table",
@@ -3667,6 +3668,7 @@ grpc_cc_library(
         "http_trace",
         "//src/core:hpack_constants",
         "//src/core:hpack_encoder_table",
+        "//src/core:metadata_compression_traits",
         "//src/core:slice",
         "//src/core:slice_buffer",
         "//src/core:time",
